+++ R/buffer_sdm.R
@@ -8,7 +8,7 @@
 #' Usually the output from \code{GBIF_data}.
 #' @param size \code{numeric}. The distance between the record and the margin of the buffer (i.e.
 #' buffer radius).
-#' @param crs \code{numeric}. Indicates which EPSG it the \code{occ_data} in.
+#' @param crs \code{numeric}. Indicates which EPSG is the \code{occ_data} in.
 #' @param mcp \code{boolean}. Should the buffer be applied in each record (FALSE) or in a minimum
 #' convex polygon/convex hull (TRUE)? Standard is \code{FALSE}.
 #'